server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
anthropic==0.2.8